M1/agents/requirements.txt
google-api-python-client
google-genai
python-dotenv

mlflow
mlflow[genai]

openai
colorama
